Create tmp directory if it doesn't already exist
GitHub's API doesn't like too many hits so systems that are hitting
refresh often without this tmp directory can quickly spam the GitHub
server and be blocked.

Warning:  This creates a directory in project/../tmp which may not be
desirable for all developers.  For example, if the site is cloned to
/home/john/lmms-website, the tmp directoy will be created in
/hom/john/tmp.